Do not flush eac3(joc) and ac4 decoders on reuse
eac3(joc) and ac4 decoders (for the same profile and level content) do
not need to be flushed to be reused for the next compatible track or
across adapative bitrate transitions. This change allows for gapless
playback on devices with Dolby decoders that require internal
re-initialization on flush.

This change introduces a codec check in canReuseCodec, in addition to
the existing mimeType check, as within the same mimeType, such as
AC-4, there can be codec changes (such as between AC-4 ac-4.02.01.04
for Atmos Objective Based Immersive and ac-4.02.02.00 for AC-4
Immersive Stereo) that may require decoder reinitialization for
correct decoding.

This is a cherry-pick of 9b1f1b1ba1269206ef12c049acf65c6299463fe9 from
https://github.com/androidx/media/pull/1346 and
https://github.com/DolbyLaboratories/media/commit/9b1f1b1ba1269206ef12c049acf65c6299463fe9
.